SBERDOMA-506 gql experiments
