+++ apps/condo/domains/ticket/gql.js
@@ -11,6 +11,10 @@ const COMMON_FIELDS = 'id dv sender v deletedAt newId createdBy { id name } upda
 const TICKET_FIELDS = `{ organization { id name } property { id name address } unitName sectionName floorName status { id name type organization { id } colors } statusReopenedCounter statusUpdatedAt statusReason number client { id name } clientName clientEmail clientPhone contact { id name } operator { id name } assignee { id name } executor { id name } watchers { id name } classifier { id name organization { id } parent { id name } } details related { id details } isEmergency isPaid meta source { id name type } sourceMeta ${COMMON_FIELDS} }`
 const Ticket = generateGqlQueries('Ticket', TICKET_FIELDS)
 
+
+const ANALITYCS_TICKET_FIELDS = '{ id createdAt property { id } status { id type } assignee { id } executor { id } }'
+const AnaliticsTicket = generateGqlQueries('Ticket', ANALITYCS_TICKET_FIELDS)
+
 const TICKET_STATUS_FIELDS = `{ organization { id } type name colors ${COMMON_FIELDS} }`
 const TicketStatus = generateGqlQueries('TicketStatus', TICKET_STATUS_FIELDS)
 
@@ -141,6 +145,7 @@ const GET_TICKET_ANALYTICS_REPORT_DATA = gql`
 
 module.exports = {
     Ticket,
+    AnaliticsTicket,
     TicketStatus,
     TicketChange,
     TicketSource,
